requirements.txt
agate==1.5.5
agate-dbf==0.2.0
agate-excel==0.2.0
agate-sql==0.5.0
appdirs==1.4.0
appnope==0.1.0
awesome-slugify==1.6.5
Babel==2.3.4
beautifulsoup4==4.5.3
bs4==0.0.1
csvkit==1.0.1
csvtools==0.5.1
dbfread==2.0.7
decorator==4.0.11
Django==1.10.5
django-extensions==1.7.5
django-filter==1.0.1
django-haystack==2.6.0
djangorestframework==3.5.3
drf-nested-routers==0.11.1
elasticsearch==2.4.1
et-xmlfile==1.0.1
future==0.16.0
html5lib==0.999999999
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
isodate==0.5.4
jdcal==1.3
leather==0.3.3
Markdown==2.6.7
openpyxl==2.4.1
packaging==16.8
parsedatetime==2.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
pytimeparse==1.1.5
pytz==2016.10
regex==2017.1.17
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.5
traitlets==4.3.1
Unidecode==0.4.20
wcwidth==0.1.7
webencodings==0.5
xlrd==1.0.0
